server/requirements.txt
pandas
scikit-learn
flask
flask-cors
requests
ipykernel
google-generativeai
dotenv